Back to old test style
